Retrofit og moshi fungerer KORREKT mot testAPI

+++ app/src/main/java/no/usn/rygleo/prisjegermobv1/data/TestAPI.kt
@@ -6,6 +6,6 @@ data class TestAPI(
     val key: String,
     val link: String,
     val participants: Int,
-    val price: Int,
+    val price: Double,  // var Int, må rettes til Double
     val type: String
 )
